% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeletonCharybdis.R
\docType{package}
\name{SkeletonCharybdis-package}
\alias{SkeletonCharybdis}
\alias{SkeletonCharybdis-package}
\title{SkeletonCharybdis: Skeleton package based on Characterizing Health Associated Risks, and Your Baseline Disease In SARS-COV-2 (CHARYBDIS)}
\description{
Skeleton package for performing large scale characterization
}
\seealso{
Useful links:
\itemize{
  \item \url{}
  \item Report bugs at \url{}
}

}
\keyword{internal}
